:recycle: split auth views.py into multiple packages
